Finished Main Models Graphql Integration + Fixed Migrations

+++ IntDesign.Calculus.Core/Models/MainModels/Room.cs
@@ -8,25 +8,25 @@ namespace Calculus.Core.Models.MainModels
     public class Room : IIdentifier
     {
         public Guid Id { get; set; }
-        public float Lenght { get; set; }
-        public float Width { get; set; }
-        public float Height { get; set; }
-        public float Pc { get; set; }
-        public float Atv { get; set; }
-        public float Asp { get; set; }
-        public float EmptyAsp { get; set; }
-        public float? Afm { get; set; }
-        public float? SpecialAfm { get; set; }
-        public float? CustomLenght { get; set; }
-        public float? CustomWidth { get; set; }
-        public float? CustomHeightOne { get; set; }
-        public float? CustomHeightTwo { get; set; }
+        public double Lenght { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
+        public double Pc { get; set; }
+        public double Atv { get; set; }
+        public double Asp { get; set; }
+        public double EmptyAsp { get; set; }
+        public double? Afm { get; set; }
+        public double? SpecialAfm { get; set; }
+        public double? CustomLenght { get; set; }
+        public double? CustomWidth { get; set; }
+        public double? CustomHeightOne { get; set; }
+        public double? CustomHeightTwo { get; set; }
         
-        public float? WallRealCoefficient { get; set; }
+        public double? WallRealCoefficient { get; set; }
         
-        public float? TilesParquetCoefficient { get; set; }
+        public double? TilesParquetCoefficient { get; set; }
         
-        public float? SpecialTilesParquetCoefficient { get; set; }
+        public double? SpecialTilesParquetCoefficient { get; set; }
         public RoomType Type { get; set; }
         public Guid HouseId { get; set; }
         public House House { get; set; }
